requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
sentence-transformers
google-genai
numpy
